requirements.txt
asyncache
async_lru
boto3
cachetools
hypercorn
markdown2
pygments
quart
quill-delta[html]